refactor: #459 post 관련 아키텍처 통일
