+++ Spoony-iOS/Spoony-iOS/Network/DependencyKey/PostServiceKey.swift
@@ -1,0 +1,12 @@
+//
+//  PostServiceKey.swift
+//  Spoony
+//
+//  Created by 최주리 on 9/25/25.
+//
+
+import Dependencies
+
+enum PostServiceKey: DependencyKey {
+    static let liveValue: PostService = DefaultPostService()
+}
